Adjust the production cdn baseURL
Probably because of an old bug in donejs-deploy we were using dist/ as
the baseURL when it should just be / (from the cdn url).  I believe the
reason this was happening was the previous version of donejs-deploy had
a bug that has been resolved. I'm going to clean out the dist/ folder on
the CDN which should no longer be needed.